% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_dashboardHeader.R
\name{sw_dashboardHeader}
\alias{sw_dashboardHeader}
\title{sw_dashboardHeader}
\usage{
sw_dashboardHeader(title = "Title", logo = NULL, add_tags = NULL)
}
\arguments{
\item{title}{string; default: Title}

\item{logo}{image name; image mast be in www subdirectory of app dir}

\item{add_tags}{additional tags for header as tagList, e.g. notification}
}
\value{
shiny object
}
\description{
sw_dashboardHeader
}
