% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_dashboardHeader.R
\name{sw_dashboardHeader}
\alias{sw_dashboardHeader}
\title{sw_dashboardHeader}
\usage{
sw_dashboardHeader(title = "Title", logo = NULL)
}
\arguments{
\item{title}{string; default: Title}

\item{logo}{image name; image mast be in www subdirectory of app dir}
}
\value{
shiny object
}
\description{
sw_dashboardHeader
}
\references{
https://github.com/rstudio/shinydashboard/issues/41
}
